bevy_transform::components
Struct GlobalTransform

pub struct GlobalTransform(/* private fields */);

GlobalTransform is an affine transformation from entity-local coordinates to worldspace coordinates.

You cannot directly mutate GlobalTransform; instead, you change an entity’s transform by manipulating its Transform, which indirectly causes Bevy to update its GlobalTransform.

To get the global transform of an entity, you should get its GlobalTransform.

For transform hierarchies to work correctly, you must have both a Transform and a GlobalTransform.

You may use the TransformBundle to guarantee this. TransformBundle is now deprecated. GlobalTransform is automatically inserted whenever Transform is inserted.

Transform and GlobalTransform

Transform transforms an entity relative to its parent’s reference frame, or relative to world space coordinates, if it doesn’t have a Parent.

GlobalTransform is managed by Bevy; it is computed by successively applying the Transform of each ancestor entity which has a Transform. This is done automatically by Bevy-internal systems in the system set TransformPropagate.

This system runs during PostUpdate. If you update the Transform of an entity in this schedule or after, you will notice a 1 frame lag before the GlobalTransform is updated.

Examples

transform

Implementations

impl GlobalTransform

pub const IDENTITY: Self

An identity GlobalTransform that maps all points in space to themselves.

pub fn compute_matrix(&self) -> Mat4

Returns the 3d affine transformation matrix as a Mat4.

pub fn affine(&self) -> Affine3A

Returns the 3d affine transformation matrix as an Affine3A.

pub fn compute_transform(&self) -> Transform

Returns the transformation as a Transform.

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

pub fn to_isometry(&self) -> Isometry3d

Returns the isometric part of the transformation as an isometry. Any scaling done by the transformation will be ignored.

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

pub fn reparented_to(&self, parent: &GlobalTransform) -> Transform

Returns the Transform self would have if it was a child of an entity with the parent GlobalTransform.

This is useful if you want to “reparent” an Entity. Say you have an entity e1 that you want to turn into a child of e2, but you want e1 to keep the same global transform, even after re-parenting. You would use:

#[derive(Component)]
struct ToReparent {
    new_parent: Entity,
}
fn reparent_system(
    mut commands: Commands,
    mut targets: Query<(&mut Transform, Entity, &GlobalTransform, &ToReparent)>,
    transforms: Query<&GlobalTransform>,
) {
    for (mut transform, entity, initial, to_reparent) in targets.iter_mut() {
        if let Ok(parent_transform) = transforms.get(to_reparent.new_parent) {
            *transform = initial.reparented_to(parent_transform);
            commands.entity(entity)
                .remove::<ToReparent>()
                .set_parent(to_reparent.new_parent);
        }
    }
}

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

pub fn to_scale_rotation_translation(&self) -> (Vec3, Quat, Vec3)

Extracts scale, rotation and translation from self.

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

pub fn right(&self) -> Dir3

Return the local right vector (X).

pub fn left(&self) -> Dir3

Return the local left vector (-X).

pub fn up(&self) -> Dir3

Return the local up vector (Y).

pub fn down(&self) -> Dir3

Return the local down vector (-Y).

pub fn back(&self) -> Dir3

Return the local back vector (Z).

pub fn forward(&self) -> Dir3

Return the local forward vector (-Z).

pub fn translation(&self) -> Vec3

Get the translation as a Vec3.

pub fn translation_vec3a(&self) -> Vec3A

Get the translation as a Vec3A.

pub fn rotation(&self) -> Quat

Get the rotation as a Quat.

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

Warning

This is calculated using to_scale_rotation_translation, meaning that you should probably use it directly if you also need translation or scale.

pub fn scale(&self) -> Vec3

Get the scale as a Vec3.

The transform is expected to be non-degenerate and without shearing, or the output will be invalid.

Some of the computations overlap with to_scale_rotation_translation, which means you should use it instead if you also need rotation.

pub fn radius_vec3a(&self, extents: Vec3A) -> f32

Get an upper bound of the radius from the given extents.

pub fn transform_point(&self, point: Vec3) -> Vec3

Transforms the given point from local space to global space, applying shear, scale, rotation and translation.

It can be used like this:

let global_transform = GlobalTransform::from_xyz(1., 2., 3.);
let local_point = Vec3::new(1., 2., 3.);
let global_point = global_transform.transform_point(local_point);
assert_eq!(global_point, Vec3::new(2., 4., 6.));

let global_point = Vec3::new(2., 4., 6.);
let global_transform = GlobalTransform::from_xyz(1., 2., 3.);
let local_point = global_transform.affine().inverse().transform_point3(global_point);
assert_eq!(local_point, Vec3::new(1., 2., 3.))

To apply shear, scale, and rotation without applying translation, different functions are available:

let global_transform = GlobalTransform::from_xyz(1., 2., 3.);
let local_direction = Vec3::new(1., 2., 3.);
let global_direction = global_transform.affine().transform_vector3(local_direction);
assert_eq!(global_direction, Vec3::new(1., 2., 3.));
let roundtripped_local_direction = global_transform.affine().inverse().transform_vector3(global_direction);
assert_eq!(roundtripped_local_direction, local_direction);

pub fn mul_transform(&self, transform: Transform) -> Self

Multiplies self with transform component by component, returning the resulting GlobalTransform