bevy_transform::systems
Function propagate_transforms

pub fn propagate_transforms(
    root_query: Query<'_, '_, (Entity, &Children, Ref<'_, Transform>, &mut GlobalTransform), Without<Parent>>,
    orphaned: RemovedComponents<'_, '_, Parent>,
    transform_query: Query<'_, '_, (Ref<'_, Transform>, &mut GlobalTransform, Option<&Children>), With<Parent>>,
    parent_query: Query<'_, '_, (Entity, Ref<'_, Parent>)>,
    orphaned_entities: Local<'_, Vec<Entity>>,
)

Update GlobalTransform component of entities based on entity hierarchy and Transform component.

Third party plugins should ensure that this is used in concert with sync_simple_transforms.