Crate bevy_transform

Bevy Transform

This crate is largely a 1:1 port from legion_transform (ecs: legion, math: nalgebra) to bevy (ecs: bevy_ecs, math: glam)

Modules

components

The basic components of the transform crate

Structs

TransformBundle

A Bundle of the Transform and GlobalTransform Components, which describe the position of an entity.

TransformPlugin

The base plugin for handling Transform components

Enums

TransformSystem

Label enum for the systems relating to transform propagation

Functions

transform_propagate_system

Update GlobalTransform component of entities based on entity hierarchy and Transform component.