Crate bevy_transform

Bevy Transform

This crate is largely a 1:1 port from legion_transform (ecs: legion, math: nalgebra) to bevy (ecs: bevy_ecs, math: glam)

Modules

commands

Extension to EntityCommands to modify bevy_hierarchy hierarchies while preserving GlobalTransform.

components

The basic components of the transform crate

helper

System parameter for computing up-to-date GlobalTransforms.

systems

Systems responsible for transform propagation

Structs

TransformBundle

A Bundle of the Transform and GlobalTransform Components, which describe the position of an entity.

TransformPlugin

The base plugin for handling Transform components

Enums

TransformSystem

Set enum for the systems relating to transform propagation

Traits

TransformPoint

A trait for point transformation methods.